requirements.txt
awscli-local
awscli
psycopg2-binary==2.9.5